refactor(backend): rename `provider` module to `traits`
