+++ apps/backend/src/audio_books/resolver.rs
@@ -10,7 +10,7 @@ use crate::{
     graphql::IdObject,
     media::resolver::{MediaSearchResults, MediaService, SearchInput},
     migrator::{AudioBookSource, MetadataLot},
-    providers::MediaProvider,
+    traits::MediaProvider,
 };
 
 use super::audible::AudibleService;
