+++ apps/backend/src/main.rs
@@ -50,8 +50,8 @@ mod media;
 mod migrator;
 mod misc;
 mod movies;
-mod providers;
 mod shows;
+mod traits;
 mod users;
 mod utils;
 mod video_games;
