+++ apps/backend/src/shows/resolver.rs
@@ -10,7 +10,7 @@ use crate::{
     graphql::IdObject,
     media::resolver::{MediaSearchResults, MediaService, SearchInput},
     migrator::{MetadataLot, ShowSource},
-    providers::MediaProvider,
+    traits::MediaProvider,
 };
 
 use super::{tmdb::TmdbService, ShowSpecifics};
