+++ apps/backend/src/video_games/resolver.rs
@@ -10,7 +10,7 @@ use crate::{
     graphql::IdObject,
     media::resolver::{MediaSearchResults, MediaService, SearchInput},
     migrator::{MetadataLot, VideoGameSource},
-    providers::MediaProvider,
+    traits::MediaProvider,
 };
 
 use super::igdb::IgdbService;
